make MVP--add players, add gameboard
This build provides all of the necessary parts to play Tic-Tac-Toe.

It does not check for winners automatically, and it does not show the player turn.

It does allow the first player to play X and the second player to play an O, while
preventing playing in an area that has already been selected.

MVP is complete. Further development can be done to improve UI/UX and re-factor some
functions for more OOP-style behavior. Also add scores and a scoreboard object, along with
an automatic check for winner or draw, and update scores from that.